+++ package.json
@@ -41,8 +41,8 @@
     "build:src": "npm run prettify && npm run lint && npm run build:fast && npm run build:doc",
     "build:fast": "gulp build && tsc",
     "build:doc": "mkdirp doc/ && ts-node --type-check -P automation/tsconfig.json automation/capitanodoc/index.ts > doc/cli.markdown",
-    "build:standalone": "ts-node --type-check -P automation/tsconfig.json automation/run.ts build:standalone",
-    "build:installer": "ts-node --type-check -P automation/tsconfig.json automation/run.ts build:installer",
+    "build:standalone": "npm run build:fast && ts-node --type-check -P automation/tsconfig.json automation/run.ts build:standalone",
+    "build:installer": "npm run build:fast && ts-node --type-check -P automation/tsconfig.json automation/run.ts build:installer",
     "release": "ts-node --type-check -P automation/tsconfig.json automation/run.ts release",
     "pretest": "npm run build",
     "test": "gulp test",
